move DAO objects to separate package
