+++ src/main/java/com/company/watches/dao/CountryDAO.java
@@ -1,5 +1,6 @@
-package com.company.watches.model;
+package com.company.watches.dao;
 
+import com.company.watches.model.Country;
 import com.company.watches.utils.DBException;
 
 import java.sql.*;
