+++ src/main/java/com/company/watches/dao/DAO.java
@@ -1,4 +1,4 @@
-package com.company.watches.model;
+package com.company.watches.dao;
 
 import com.company.watches.utils.DBException;
