+++ src/main/java/com/company/watches/dao/DAOContainer.java
@@ -1,4 +1,6 @@
-package com.company.watches.model;
+package com.company.watches.dao;
+
+import com.company.watches.model.*;
 
 public class DAOContainer {
