+++ src/main/java/com/company/watches/manage/VendorManager.java
@@ -1,8 +1,8 @@
 package com.company.watches.manage;
 
+import com.company.watches.dao.DAO;
+import com.company.watches.dao.DAOContainer;
 import com.company.watches.model.Country;
-import com.company.watches.model.DAO;
-import com.company.watches.model.DAOContainer;
 import com.company.watches.model.Vendor;
 import com.company.watches.utils.DBException;
